feat: initial commit for the new multi-step form
