+++ src/app/deals/add/page.tsx
@@ -1,5 +1,5 @@
+import Separator from '@/components/Separator'
 import DealForm from './DealForm'
-import { redirect } from 'next/navigation'
 
 //TODO: why doesn't this work with server component page?
 // export const metadata: Metadata = {
@@ -8,15 +8,15 @@ import { redirect } from 'next/navigation'
 // };
 
 export default function AddDealPage() {
-  redirect('/')
   return (
     <main className="">
-      <h1 className="py-20 text-center text-4xl font-bold text-gray-100">
-        Share a Deal
-      </h1>
-      <div className="mx-auto max-w-2xl pb-20">
-        <DealForm />
-      </div>
+      <h1 className="mb-4 text-7xl font-semibold text-white">Share a Deal</h1>
+      <span className="text-2xl font-light text-white">
+        Have an amazing deal or discount tailored for developers? Let us know!
+      </span>
+      <Separator className="mb-32 mt-14" />
+
+      <DealForm />
     </main>
   )
 }
