+++ src/components/forms/CategorySelect.tsx
@@ -4,29 +4,32 @@ import {
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from '@/components/ui/select';
-import { Category } from '@/types/Types';
+} from '@/components/ui/select'
+import { Category } from '@/types/Types'
 // import { SelectLabel } from '@radix-ui/react-select';
 
 export default function CategorySelect({
   onCategoryChange,
 }: {
-  onCategoryChange: (category: Category) => void;
+  onCategoryChange: (category: Category) => void
 }) {
   return (
-    <Select onValueChange={onCategoryChange}>
-      <SelectTrigger className="w-[300px]">
-        <SelectValue placeholder="Course" />
-      </SelectTrigger>
-      <SelectContent>
-        {Object.values(Category)
-          .sort()
-          .map((dealType) => (
-            <SelectItem key={dealType} value={dealType}>
-              {dealType}
-            </SelectItem>
-          ))}
-      </SelectContent>
-    </Select>
-  );
+    <div className="flex flex-col gap-2">
+      <span className="text-2xl font-extralight">Category *</span>
+      <Select onValueChange={onCategoryChange}>
+        <SelectTrigger className="h-16 w-[334px] bg-transparent">
+          <SelectValue placeholder="Course" />
+        </SelectTrigger>
+        <SelectContent>
+          {Object.values(Category)
+            .sort()
+            .map((dealType) => (
+              <SelectItem key={dealType} value={dealType}>
+                {dealType}
+              </SelectItem>
+            ))}
+        </SelectContent>
+      </Select>
+    </div>
+  )
 }
